chore: split samples into individual files
